Fix problems with buffers filling up in format_ndag.c
If there is no data available for reading on the fd, we still
need to treat the stream as "ready" for reading if there are any
filled buffers. Failure to do so can lead to incorrectly assuming
that all sockets are inactive, which will put the thread to sleep.

Also added a slight performance enhancement where we avoid
calling select() if all of the buffers are nearly full.