Improve error explanation when in-libtrace filtering fails.
The typical failure case is because the packet linktype does not
correspond to a pcap DLT, which BPF needs to apply filters
properly.

Since the packets being filtered may not actually be pcap, an
error message that complains about pcap is rather confusing.
Better to be more detailed about what went wrong.