Return error if change perpkt threads on paused trace
We do not have code to change the number of perpkt threads, once
a trace had been started. However, you could set the perpkt threads
without an error being returned on a paused trace - but this never
got applied.

This clarifies docs and returns an error to match with the behaviour.