Fix bug where two ring: trace inputs had same fanout group.
We use rand() to assign the fanout group when we create the
input, but the rand() seed state is usually equal in both threads
at the point where we call rand().

So we end up with two inputs trying to be assigned to the same
fanout group, which fails and the second input will immediately
halt.

Also added code to try incrementing the fanout group number if
we end up clashing with another existing group. This will also
resolve any issues with the RNG producing one-off clashes.