Replace AC_CHECK_FUNCS with AC_CHECK_DECLS
It seems that `AC_CHECK_FUNCS` does not correctly detect when these "functions" are implemented as macros. (I've seen this happening on macOS, and Ubuntu 16.04 with `snprintf` and `strndup` respectively.) Replacing `AC_CHECK_FUNCS` (and the similar `AC_REPLACE_FUNCS`) with `AC_CHECK_DECLS` seems to solve this problem, though I don't have access to any systems that do not have these functions to be sure that it correctly detects them as missing.