Don't munmap ring rx buffers until last possible moment.
As soon as we munmap the buffers, any packets contained within
those buffers become invalid -- however, the user application may
still have references to those packets and could try to operate
on them without knowing that the memory holding the packet payload
is invalid.

Previously, the buffers were munmapped when the trace was paused,
which was causing crashes and invalid memory accesses.

Instead, we now only munmap the buffer if either the trace is
restarted or destroyed, so the packets will remain valid for
longer. Also, the new "start" iteration counting added in an
earlier commit means we can now recognise when a packet belongs
to an unmapped buffer and therefore return appropriate errors if
a user tries to interact with the packet.

This also fixes the assertion failure when freeing a ring packet
after the ring trace has been paused, i.e. after a call to
trace_pstop().