ndag: join the stream socket multicast group at the last minute
Joining earlier means that we have packets queuing up while we're
still allocating buffers etc. This means we run the risk of
dropping packets during our init phase.

Instead, we now do as much initialisation as possible before
joining the group -- it means that our first packet might end up
being a later one than we would have read previously, but we are
less likely to have a gap in our incoming stream.

Using the old method, our first batch of packets might have been
as follows:
  1, 2, 3, 4, 5, 8, 9, 10

Now we might miss 1 and 2, but avoid having the gap from 5 to 8,
i.e.:
  3, 4, 5, 6, 7, 8, 9, 10

I think generally the latter is a better sequence of packets for
us to provide to the user. It also means that our reported missing
count is zero, which is less likely to make users think that there
is a performance issue in their own code.