ndag: just read one "batch" of packets per pread call
Previously, we would stay in pread until we got close to the
full number of packets requested or a non-blocking read gave us
no new packets.

This meant that we could in theory waste time doing an unnecessary
non-blocking read that was going to fail, when instead we could
just assume that we were going to be called again shortly anyway
if the user wanted more packets.

It's a bit tenuous when I try to explain it, but it seems to save
us a bit of CPU time in practice...