Add option to allow users to specify a constant ERF framing length.
This can be useful for situations where
  a) the input uses ERF (or some derivative thereof)
  b) the link type and ERF record type are constant for the
     duration of the capture
  c) performance is critical

This option allows users to simply tell libtrace what the ERF
framing length on every packet is going to be, so libtrace
doesn't have to repeatedly derive the framing length for each
packet it processes. At high packet rates, the time taken to
do this calculation can really add up and it makes no sense
to risk dropping packets because you're busy calculating a value
that is always a single constant value.