Prefer an installed .so over a .a for libdpdk.
We sometimes run into issues with building the libtrace tools
if we link against the .a (e.g. if we've installed dpdk-dev on
Debian jessie).

I think we're getting close to the point where we can consider
ignoring the .a altogether -- the .so is installed by default
by the Debian packages and our own DPDK instructions tell
people how to build the shared library -- so maybe we can
reasonably expect that there will be a .so available if we need
it and if not, then we can just build without dpdk support?