Don't get unhappy if a pcap packet has a wire length of 65534
Our internal wire length calculation will add 4 bytes to the
pcap wire length, which means it is possible (albeit incredibly
unlikely) to have a pcap packet with a libtrace wire length >
LIBTRACE_PACKET_BUFSIZE (e.g. 65534 + 4 bytes > 65536).

Therefore, if we are a pcap packet, add an extra four bytes of
leeway when determining if a wire length is bogus or not.

Bug discovered by Perry using the afl fuzzer.