Add test case for trace_get_wire_length()
Given that there is often some confusion around the wire length
and whether we count the FCS as wire length or not (and what do
we do for formats that don't save the FCS, i.e. pcap), we should
really have a test that will at least alert us if someone has
changed how wire length is calculated for one of the formats.